Added Preferences, fixed naming, adjusted layout
At long last I gave this one a little love, linked all the most important values to user adjustable settings, so, if you're always adjusting the mesh size, or light intensity, you can now set them as user defaults in the Addon Preferences menu.
The repo was originally called Unreal Objects, it's been renamed to Custom Objects, and the naming convention has been updated and enforced throughout the code.
Originally had the 3D Viewport Panel contained within an invisible Parent Panel to give the user the option to hide or show it easily. This has been replaced with a simpler layout, eliminating the invisible box, and replicating the functionality with a 'Show In Viewport' boolean, also in the Addon Preferences.

Next steps would be to break it up from one file into multiple smaller files to allow for easier customization by the end user, and documenting how to add your own custom objects.